separate tools from application go.mod
it's been great to keep track of the tools we need and their specific
versions by tracking them with `go.mod`, but it turns out that there are
tons of deps that get unecessarily introduced into our `go.mod` that
would be checked for, say, licenses if we were to scan out repository.

by breaking those go.mod apart, we can then proceed with scans solely on
the main `go.mod` and ignore the tools deps.

Signed-off-by: Ciro S. Costa <ciroscosta@vmware.com>