Add ytt support for templates
In addition to the current tempesting scheme using JSON Path,
Cartographer now support a sibling scheme that uses ytt. ytt has rich
support for applying data values to structured yaml with the full
expressiveness of a programming language that is executed within a
sandbox.

As a sibling to each Template field on the components is a Ytt field.
These fields are mutually exclusive. The ytt template is passed resolved
with the template context values bound as data values. For example, with
SupplyChain processing the Workload fields are available within ytt at
`data.values.workload`.

The ytt binary is added to the container by ko. At the moment only
linux-amd64 is supported.

Signed-off-by: Scott Andrews <andrewssc@vmware.com>