hack/ci: remove scripts that are not ci specific
some of the files we had under `hack/ci` were not really just CI
specific - they could be run locally, and have no dependency on a CI at
all, like `e2e.sh` which has been very useful for running the end to end
example that we have to completion even locally.

aside from the removals, this commit adds the `install-binaries` script
that allows us to, with a call to it, install binaries that are relavant
to CI, and assumes that it is running in a CI environment (e.g., see use
of `sudo`).

usage:
	./hack/ci/install-binaries.sh <dependency>

e.g., to install `ko` and `kuttl`:

	./hack/ci/install-binaries.sh ko kuttl

(if you're wondering why make it so that it takes args: this way we can
look at a github workflow and not exactly what is being installed, as
well as let us not install things we don't need in certain workflows,
like how a `validation` workflow would certainly not need the `gh` cli).

Signed-off-by: Ciro S. Costa <ciroscosta@vmware.com>