hack: add release generation script
with the introduction of carvel packaging[1], there'd be more complexity
added to the `make release` target that I believe is better to be
introduced to a bash script of its own rather than the Makefile.

this script then takes care of generating the contents of the `/release`
directory (gitignored) by preparing an `imgpkg` bundle[2] containing all
the relevant files for installing cartographer, as well as pushing that
bundle and then templating out the final package.

in the end, we have the following structure generated:

	./release/
	|
	├── cartographer.yaml		- installation for a `k apply -f-`
	|
	└── package			- installation via carvel pkging
	    ├── package-install.yaml
	    ├── package-metadata.yaml
	    └── package.yaml

[1]: https://carvel.dev/kapp-controller/docs/latest/packaging/
[2]: https://carvel.dev/imgpkg/docs/latest/

Signed-off-by: Ciro S. Costa <ciroscosta@vmware.com>