hack: add github-release script
this script allows us to publish a github release, be on CI or not. note
that this does not generate any of the files related to the release - it
assumes that the files are already there, and so it just makes use of
github's cli[1] to submitting the release through its api. see
`hack/release.sh` and `hack/release-notes.sh` for release assets
generaiton.

example: use the current tag as the release version (default), the
`release/` directory as the place where `yaml` files to be released can
be found (default), and `release/CHANGELOG.md` as the release notes
body (default):

	./hack/github-release.sh

see `readonly` vars in the script for customization.

[1]: https://github.com/cli/cli

Signed-off-by: Ciro S. Costa <ciroscosta@vmware.com>