hack: re-introduce cluster setup
this commit brings back the functionality that we had for the previous
`hack/e2e.sh` script, but:

- broken down into separate commands
- by default, making use of an authenticated registry rather than not
- removing all resource reservations from our dependencies
- leveraging secretgen[1] for passing registry credential secrets to all
  namespaces
- using `kapp-ctrl/Package*` as the way of installing cartographer

with this setup we end up then replicating the setups of the real-world
much more closely, letting us be aware of the tweaks necessary for
working with private registries.

by having the funcionality of e2e.sh broken down into different
commands, it also becomes easier to bring a cluster locally and iterate
on an example, of install _just_ a subset of the dependencies you want.

for instance:

	1. bring up a local cluster with a registry and base
	dependencies we'd expect to be there
	(kapp-ctrl,cert-manager,secretgen):

		./hack/setup.sh cluster

	2. install cartographer using local code, pushing that to the
	local registry

		./hack/setup.sh cartographer

	3. setup the extra dependencies our examples use (like kpack,
	knative, etc)

		./hack/setup.sh example-dependencies

	4. run the example end to end

		./hack/setup.sh example

[1]: https://github.com/vmware-tanzu/carvel-secretgen-controller

Signed-off-by: Ciro S. Costa <ciroscosta@vmware.com>