hack: make bin installs uniform
kubebuilder didn't follow the same pattern of having all of the binaries
landing on /usr/local/bin, which by default is open to our non-root
user on ubuntu.

by switching kubebuilder bins to that direcory, we have to let envtest
know that that's where it can find those binaries (apparently, it
doesn't search $PATH by default).

Signed-off-by: Ciro S. Costa <ciroscosta@vmware.com>