hack: turn docker-login resilient to bad json
previously, we'd only guard against the case where the DOCKER_CONFIG
(usually ~/.docker/config.json) file is either not present or is
0-sized, but the best is to actually make sure that we have there is
good for appending an auth section to the registry we want to add
credentials for.

by optimistically performing a `jq '.'` we can move on to create an
empty `{}` (valid json) whether the file exists or not, as well as
whether the file contains a valid json or not.

Signed-off-by: Ciro S. Costa <ciroscosta@vmware.com>