hack: improve resource removal overlay
the previous form of overlay that we had was very specific to matching
only deployments that were already making use of `resources`.

it turns out that `tekton` doesn't, so it wouldn't be covered.

given that it's also very likely that there might be other gvrs that
we'd want to cover (like deployment, daemonsets and statefulsets), this
takes care of all of that.

Signed-off-by: Ciro S. Costa <ciroscosta@vmware.com>