docs: update installation instructions
given that this repository now is open, it makes sense to provide
instructions for how to install cartographer based on images that we
publish to dockerhub.

the approach I took here was to keep the other usecases (which are ..
good usecases), but immediatelly provide the instructions that folks
would expect: `k apply -f <release>`.

Signed-off-by: Ciro S. Costa <ciroscosta@vmware.com>