hack: make release produce cartographer.yaml
given that we want to allow folks to install cartographer _without_
having to go through the trouble of using relocation to do so, it makes
sense to have to expose the single `cartographer.yaml` file that gets
produced during the bundle generation and making that available.

note that we don't have to change anything else w/ regards to changelog
etc because this will already be picked up by just existing inside that
directory.

e.g.:

	```
	bf3f222c57f3d2939a37c438f9002dbee6620ad28a415e535b50fe6b43a14c97  ./package/package-metadata.yaml
	8dd3734081c38e699c73c7bb6706324b3c8cdda1f6bc1cc9c7a0a70895575b58  ./package/package-install.yaml
	3284491441d2b57b2790dd50fbea52370e1ee8acb27f9be96f7d089fc14354ef  ./package/package.yaml
	a8584490f15aa647da8d3c0073608b5721e9a774b9fd7380bb4daa95898a1819  ./bundle.tar
	80cde5c816b585cf776899c52ed3bb9f47a953ce5da163c40daa735f7df2f96a  ./cartographer.yaml
	```

Signed-off-by: Ciro S. Costa <ciroscosta@vmware.com>