Blueprint params specify default or value
Params on supply-chains and deliveries (and on the associated
resources) require either a `default` or `value`.

If the `default` field is specified, the provided value can be overriden
by the param supplied by the workload.

If instead the `value` field is specified, the param value supplied
by the workload is ignored.

If for the same named param, the blueprint specifies a value at
the top level and at an individual resource level, the specification
at the individual resource level will clobber the specification at the
top level.

[#354]

Authored-by: Waciuma Wanjohi <lwanjohi@pivotal.io>