examples: add note about doc versioning
as we update the examples with the latest features, it'll be inevitable
that our releases will not be ready yet and people will go throguh the
example expecting it to "just work".

given that "just work" will always be true for examples under a
particular tag (e.g., v0.0.7), here I point folks at such.

Signed-off-by: Ciro S. Costa <ciroscosta@vmware.com>