examples: use of single serviceaccount
at the moment, we make use of `serviceAccount` in some places referring
to the `service-account` SA (which contains secrets for reaching out to
the registry w/ authz granted), but in others, we use the `default` one
(relying on people having set that up appropriate according to the
documentation).

this commit changes that, moving toward a more explicit use of a single
serviceaccount (the one named `service-account` that we create).

this should contribute to reducing some implicit problems that might
occur, like how with the lack of `serviceAccount` in the revision
templatespec for knative we end up seeing problems if `default` points
at a secret that doesn't exist anymore (as knative tries to fetch
secrets from the `default` SA if not specified)

Signed-off-by: Ciro S. Costa <ciroscosta@vmware.com>