hack: remove carvel-based packaging
with the move towards focusing packaging-related activities under
`vmware-tanzu/package-for-cartographer`, we can remove some of the
complexity of releasing that we had here.

with this commit, rather than always create a release directory that
ends up in `cartographer.yaml` and `Package`-related files (backed by an
imgpkg bundle), we now only need to take care of creating the
`cartographer.yaml` (using `ko`) and .. that's it.

Signed-off-by: Ciro S. Costa <ciroscosta@vmware.com>