config: add `app.kubernetes.io/component` label
following the best practices when it comes to labelling objects (see
[1]), by making use of the recommended `app.kubernetes.io/component`,
we're able to clearly identify those objects that are part of
cartographer (useful for any tool that might want to do some form of
matching against anything cartographer-related, e.g., getting rid of all
of them based on matching on such label).

the approach I took here is to make use of ytt overlays in order to add
the label to all components _except_ the Namespace.

[1]: https://kubernetes.io/docs/concepts/overview/working-with-objects/common-labels/

Signed-off-by: Ciro S. Costa <ciroscosta@vmware.com>