hack: override default validatingwebhook name
by default, controller-tools' controller-gen generates the object with a
hardcoded name of `validating-webhook-configuration`

for those using kustomize (like kubebuilder does), that's not a problem
as the configuration usually specifies what the prefix for objects
should look like, e.g.

	# in kustomization.yaml

	namePrefix: cartographer-conventions-

but, given that this repository _doens't_ make use of kustomize (as it
wasn't initialized w/ kubebuilder), here we resort to using ytt overlays
for it (we already had one for webhook configs anyway, so, just update
it).

see
https://github.com/kubernetes-sigs/controller-tools/blob/ebdf5177ad92f89da82686f4100c06d428ab3ce5/pkg/webhook/parser.go#L384-L405
for details on that hardcoded name.

Signed-off-by: Ciro S. Costa <ciroscosta@vmware.com>